requirements.txt
# Libraries

python-snap7